requirements.txt
python-dotenv
pandas
psycopg2-binary
matplotlib
SQLAlchemy
requests
alpha_vantage
apache-airflow